+++ RELEASE_NOTES.md
@@ -13,6 +13,8 @@
 - Allow multiplying and dividing any `Quantity` by a `float`. This just scales the `Quantity` value.
 - Allow dividing any `Quantity` by another quaintity of the same type. This just returns a ration between both quantities.
 
+- The battery pool `power` method now supports scenarios where one or more inverters can have multiple batteries connected to it and one or more batteries can have multiple inverters connected to it.
+
 ## Bug Fixes
 
 - Fix grid current formula generator to add the operator `+` to the engine only when the component category is handled.